+++ packages/apptive_grid_core/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.12.2
 * Add `hiddenFields` to `Grid`
 * Deprecate `ApptiveLinkType.rename` in favor of `ApptiveLinkType.patch`
+* Adjust Format for equals Filter
 
 ## 0.12.1
 * Make `defaultHeaders` in ApptiveGridClient public outside of tests